requirements.txt
python-telegram-bot==20.7
Pillow==10.4.0
